Refactored data preprocessing
- Moved data preprocessing functions into separate module
